api/requirements.txt
flask
python-dotenv
python-jose
flask-cors
six
gunicorn
boto3
uuid